requirements.txt
lxml==4.9.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.21.5
opencv_python_headless==4.7.0.68
ortools==9.5.2237
pandas==1.4.4
pytest==7.1.2
pytest_cases==3.6.13
PyYAML==6.0
scipy==1.9.1
soupsieve==2.3.1
